+++ api/py/tract/__init__.py
@@ -49,7 +49,7 @@ from .model import Model
 from .inference_model import InferenceModel
 from .runnable import Runnable
 from .runtime import Runtime, runtime_for_name
-from .transform import TransformSpec, ConcretizeSymbols, FloatPrecision, Pulse
+from .transform import TransformSpec, SetSymbols, FloatPrecision, Pulse
 from .nnef import Nnef
 from .onnx import Onnx