+++ core/src/ops/array/tile.rs
@@ -45,7 +45,7 @@ impl EvalOp for Tile {
 impl TypedOp for Tile {
     as_op!();
 
-    fn substitute_symbols(
+    fn set_symbols(
         &self,
         _source: &TypedModel,
         node: &TypedNode,