+++ core/src/ops/source.rs
@@ -66,7 +66,7 @@ impl TypedOp for TypedSource {
         )))
     }
 
-    fn substitute_symbols(
+    fn set_symbols(
         &self,
         _source: &TypedModel,
         node: &TypedNode,