+++ cuda/src/ops/quant_q81.rs
@@ -104,7 +104,7 @@ impl TypedOp for CudaGgmlQuantQ81 {
         .with_context(|| format!("Error while computing facts for {:?}", self.name()))
     }
 
-    fn substitute_symbols(
+    fn set_symbols(
         &self,
         _source: &TypedModel,
         node: &TypedNode,